Issue: React to a winning guess #6
When the player has won the game, a new stage appears. The new stage features a GIF congratulating the player. Moreover 2 buttons appear. The player can select between Play Again or Quit the game.

The implementation of the source code can be found in Congratulations.java.

For testing purposes, a button has been created in Game.java to check the proper functions of the Congratulations Stage.
For reference: Line 74-78 & 111-118 in Game.java