Issue #3: Try to find a way to overlay 2 images
Everything is working now. Serveral changes have been made, which I'll explain briefly below.

Looking in CrossOut.java. Line: 23 - 26.  The technique to overlay 2 images is by using the BlendMode method.
Line 23-26 creates a new Group called blend, which overlays the bottom and the top imageview.

Thing is when I tried to add the blend (object?) in the imageView Arraylist<ImageView> in Display.java, I get an error message due to the fact that the blend object is not an ImageViewObject.

So in order to avoid this problem, I changed the loadPics() method (Line 65 - 76) in Display.java. So instead of adding the ImageViews in an Arraylist<ImageView>, I immediately put the ImageViews into an HBox and stored them in an Arraylist<HBox> called imageViewsHBox. By doing this, I have no trouble to add the blend object into the Arraylist<HBox> imageViewsHBox.

Holy shit. I'm bad at writing hahah. I'll explain it to you once again in detail if you don't get my message above.

Anyway (Line 78 - 87), we save the paths of the images in a ArrayList<String>.
This is useful the crossOutPic() method (Line 101 - 105). First we look for the index of the ImageView to replace, then replace it using imageViewsHBox.set(index, crossOut.getLayout()); (Line 104)

For tesing purposes, look at Line 58-59. There we replace the Maria.png